refactor(cache-schema): organize std schema impls
